pyproject.toml
[tool.poetry]
name = "python-kafka-cc-frauddetection-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.3"
scikit-learn = "^1.5.2"
joblib = "^1.4.2"
kafka-python = {git = "https://github.com/dpkp/kafka-python.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
